stabilize code for mkdna
ibmp, xbmp, nabla bitmap are introduced.

+++ nabla_dna.h
@@ -1,11 +1,13 @@
 #ifndef _NABLA_DNA_H_
 #define _NABLA_DNA_H_
 
+#include <stdint.h>
+
 #include "ibmp.h"
 
 typedef struct {
 	int	size;
-	int	*pixels;
+	uint8_t	*pixels;
 } dnabla_t;
 
 extern dnabla_t *build_nabla_dna(ibmp_t *ibmp, int dna_size);
